Quick and dirty factory class to allow speccing controllers
In your controllers, rather than View::factory and friends, provide
a way to inject a ViewFactory (or create a new one as an empty
default). Create your views as $this->view_factory()->create().

This then means you can inject a stub ViewFactory in order to stub
view classes and test expectations about the view data that will be
set.

You can also extend the ViewFactory class to map class names to
alternates if required.