Add README and LICENCE
The status of the licence is not 100% clear. The previous versions - of which this follows the same
broad concept but does not really share any code - were released open source without specific
licence files and in some cases with generic copyright headers that do not identify the copyright
holder. I've attempted to acknowledge prior contributions in the readme and licence.