feat: Drop PageContentView and PageLayoutView for generic alternatives
PageContentView and PageLayoutView were the original implementation of
supporting nested views, presuming that content would always be wrapped
in exactly one outer layout. This later proved not to be the case and we
introduced the more flexible NestedChildView and NestedParentView to
support cases where a view might have more than one parent (e.g. an
"inner layout" with a sidebar & content area, contained in the overall
site layout).

At the time, we documented that PageLayoutView and PageContentView were
deprecated and would be removed (although they were not phpdoc tagged as
such).

On review, some of the type signatures between the different variations
- which were only in phpdoc - are actually incompatible. For example,
`AbstractPageContentView` was typed as taking `PageLayoutView` as the
parent, but it implements `NestedChildView` which specifies that
`getParentView()` will return a `NestedParentView` - a `PageLayoutView`
was not compatible with that signature.

It's therefore impossible to keep the existing structure with strict
types.

Dropping the old interfaces also gives more flexibility in end-user code
- a child view will now have to accept any `NestedParentView` rather
than explicitly a `PageLayoutView`. If the layout is later reorganised
to introduce an intermediate view, that will not require any change to
the child views.