feat: Don't fail on complex / virtual types
PhpStan treats these as IntersectionType and then turns that into a
null for the php-parser type (because they are not intersections
of objects). That then causes a failure trying to build a new property
without a type.

Cast everything to `mixed` at least for now, there might be a way to
find the actual base type (e.g. array) but it could prove complex.