fix: Make validating view vars internal to the AbstractViewModel
There should be very limited cases where users now need to actually
customise the validation of display variables - type safety is covered
by PHP itself and the `ViewModelProperty` attribute is available to
control whether a property can be provided to display / should have a
default value.

Remove the old method to ensure that end-users are forced to update
their code.