feat: Make `raw()` an explicit function & move escaping responsibility
Previously, `raw()` never existed as an actual function, instead it was
a signal to the TemplateCompiler not to wrap the output in an escaping
function.

We then started to provide a function stub (that was never actually
loaded) to remove IDE etc warnings on calls to an undefined function.
However, the function stub threw an exception to guard against an
end-user accidentally including & calling it.

Improved IDE and static analysis tooling means that the first `raw` call
in a template is detected as `return: never`. This causes warnings about
"dead code" in the remainder of the template, and can cause issues with
usage detection etc.

Instead, we make `raw` an actual runtime function that can explicitly
mark variables as being safe to render unescaped. Defining this as an
interface & DTO also allows applications to generate these within
ViewModels etc, removing responsibility from the template to decide what
should and should not be rendered `raw`.

To implement this cleanly means adding a layer around the native kohana
HTML::chars() escaping function. The logical place to do this is to move
it into the Renderer. It is then no longer necessary to allow
customising the escaping method for the TemplateCompiler - the user can
simply implement & register a custom Renderer implementation.

This does slightly increase the nesting & number of function calls at
runtime but with modern PHP the performance implication is expected to
be inconsequential compared to the improved DX.