feat: Use distinct DisplayVariableAttribute instead of ViewModelProperty (#39)
When previewing an app migration, we found that the use of
`ViewModelProperty` was not that clear. In particular because in a model
like this:

```php
class MyView {
  #[ViewModelProperty(is_displayable: false)]
  public string $foo;
}
```

There was potential to think that the `->foo` prop would somehow not be
available to display in the template, rather than this meaning it could
not be passed to `>display()`.

We also realised that it was possible (but invalid / confusing) to do
`#[ViewModelProperty(is_displayable: false, is_optional: true)]`.

In fact there are only 3 valid states for a property and they are
mutually exclusive:

* It *must* be passed to ->display()
* It *can* be passed to ->display(), and will reset to default if not
* It *must not* be passed to display().

This is better expressed as three separate attributes. I have based them
on an interface for simplicity in finding them - this theoretically
would also allow for custom attributes to implement that interface,
although I don't think at present there's any valid case for doing so.