fix: Allow `raw()` to accept null values (#42)
* fix: Allow `raw()` to accept null values

Otherwise any old view that is using it would need to be reviewed and
updated to do a `?? ''` within the viewmodel / template. Simpler to
just let it coalesce to empty string within the raw function itself.

* fix: Implement __toString on UnescapedHtmlSafeString.php

Avoids IDE and static analysis warnings when using syntax like
`<?=raw('some string');?>` in a template, by showing that the object
returned from raw will be stringable.