feat: Implement v5 migration tool to convert views to native properties (#38)
The migration tool provides a set of Rector rules and configurations to 
perform the initial migration of a project to the new 5.x interfaces,
model definition API and template syntax.

This will significantly reduce the "grunt work" required to migrate
a project, but the output will need to be very carefully reviewed and
it is highly likely some manual updates will still be required.