feat: Refactor parsing display variable schemas & support optional cache (#45)
Extract the responsibility for parsing an `AbstractViewModel` display
variable schema from the `AbstractViewModel` itself, in favour of a
global instance of `ViewModelDisplaySchemaProvider`. This allows for
custom parsing strategies, if required.

The global `ViewDisplaySchemaProviderInstance` will use the default
provider with no caching by default, but can be initialised with a
custom provider and/or a caching layer. This is highly recommended for
runtime use.